requirements.txt
streamlit
yfinance
finnhub-python
pandas
ta
plotly
numpy